+++ src/main/java/com/prgrms/needit/domain/user/email/repository/EmailCodeRepository.java
@@ -1,5 +1,6 @@
-package com.prgrms.needit.common.email;
+package com.prgrms.needit.domain.user.email.repository;
 
+import com.prgrms.needit.domain.user.email.entity.EmailCode;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
